pyproject.toml
[project]
name = "fastapi-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi[standard]>=0.128.0",
    "greenlet>=3.3.1",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "sqlalchemy>=2.0.46",
]
